Changed GameIds/AccountIds/SummonerIds should be Int64s to prevent overflowing.

+++ LoLNotes/Messages/Account/AccountSummary.cs
@@ -41,6 +41,6 @@ namespace LoLNotes.Messages.Account
 		public string Username { get; set; }
 
 		[InternalName("accountId")]
-		public int AccountId { get; set; }
+		public Int64 AccountId { get; set; }
 	}
 }
